restructure database folder and create script to start honeypot
