+++ honeypot/src/main/database/config_temp/TempConfigObject.py
