+++ honeypot/src/main/database/config_temp/plugins.cfg
