+++ honeypot/src/main/main.py
@@ -1,0 +1,4 @@
+from framework.frmwork import Framework
+
+frame = Framework('/config/plugins.cfg')
+frame.start()
